drop the numpy reference implementation from the shipped tree
btgrid_numpy was only consumed at runtime through load_btgrid_shards;
move that into btgrid_cache and keep the numpy reference + parity tests
in the development tree only.
